+++ rr/soa.js
@@ -7,38 +7,43 @@ class SOA extends RR {
 
     // name is the zone name
 
-    // ttl, minimum (used for negative caching, since RFC 2308)
-    // RFC 1912 sugggests 1-5 days
-    // RIPE recommends 3600 (1 hour)
-
-    const fields = [ 'mname', 'rname', 'serial', 'refresh', 'retry', 'expire' ]
+    const fields = [ 'minimum', 'mname', 'rname', 'serial', 'refresh', 'retry', 'expire' ]
     for (const f of fields) {
-      this[f](opts[f])
+      this[`set${f.charAt(0).toUpperCase() + f.slice(1)}`](opts[f])
     }
   }
 
+  // minimum (used for negative caching, since RFC 2308)
+  // RFC 1912 sugggests 1-5 days
+  // RIPE recommends 3600 (1 hour)
+  setMinimum (val) {
+    if (!this.is32bitInt('SOA', 'minimum', val)) return
+
+    this.set('minimum', val)
+  }
+
   // MNAME (primary NS)
-  mname (val) {
+  setMname (val) {
     if (!this.validHostname('SOA', 'MNAME', val)) return
     if (!this.fullyQualified('SOA', 'MNAME', val)) return
     this.set('mname', val)
   }
 
   // RNAME (email of admin)  (escape . with \)
-  rname (val) {
+  setRname (val) {
     if (!this.validHostname('SOA', 'RNAME', val)) return
     if (!this.fullyQualified('SOA', 'RNAME', val)) return
     if (/@/.test(val)) throw new Error('SOA RNAME replaces @ with a . (dot).')
     this.set('rname', val)
   }
 
-  serial (val) {
+  setSerial (val) {
     if (!this.is32bitInt('SOA', 'serial', val)) return
 
     this.set('serial', val)
   }
 
-  refresh (val) {
+  setRefresh (val) {
     // refresh (seconds after which to check with master for update)
     // RFC 1912 suggests 20 min to 12 hours
     // RIPE recommends 86400 (24 hours)
@@ -47,7 +52,7 @@ class SOA extends RR {
     this.set('refresh', val)
   }
 
-  retry (val) {
+  setRetry (val) {
     // seconds after which to retry serial # update
     // RIPE recommends 7200 seconds (2 hours)
 
@@ -56,13 +61,27 @@ class SOA extends RR {
     this.set('retry', val)
   }
 
-  expire (val) {
+  setExpire (val) {
     // seconds after which secondary should drop zone if no master response
     // RFC 1912 suggests 2-4 weeks
     // RIPE suggests 3600000 (1,000 hours, 6 weeks)
     if (!this.is32bitInt('SOA', 'serial', val)) return
 
     this.set('expire', val)
+  }
+
+  toBind () {
+    return `$TTL    ${this.get('ttl')}\n$ORIGIN ${this.get('name')}.\n${this.get('name')}.   ${this.get('class')}  SOA ${this.get('mname')}    ${this.get('rname')} (
+          ${this.get('serial')}
+          ${this.get('refresh')}
+          ${this.get('retry')}
+          ${this.get('expire')}
+          ${this.get('minimum')}
+          )\n\n`
+  }
+
+  toTinydns () {
+    return `Z${this.get('name')}:${this.get('mname')}:${this.get('rname')}:${this.get('serial')}:${this.get('refresh')}:${this.get('retry')}:${this.get('expire')}:${this.get('minimum')}:${this.get('ttl')}:${this.get('timestamp')}:${this.get('location')}\n`
   }
 }