+++ test/soa.js
@@ -10,12 +10,13 @@ const validRecords = [
     class  : 'IN',
     name   : 'example.com',
     type   : 'SOA',
-    mname  : 'matt.example.com.',
-    rname  : 'ns1.example.com.',
+    mname  : 'ns1.example.com.',
+    rname  : 'matt.example.com.',
     serial : 1,
     refresh: 7200,
     retry  : 3600,
     expire : 1209600,
+    minimum: 3600,
     ttl    : 3600,
   },
 ]
@@ -25,8 +26,8 @@ const invalidRecords = [
     class  : 'IN',
     name   : 'example.com',
     type   : 'SOA',
-    mname  : 'matt.example.com.',
-    rname  : 'ns1.example.com.',
+    mname  : 'ns1.example.com.',
+    rname  : 'matt.example.com.',
     serial : 4294967296,
     refresh: 7200,
     retry  : 3600,
@@ -38,4 +39,26 @@ const invalidRecords = [
 describe('SOA record', function () {
   base.valid(SOA, validRecords)
   base.invalid(SOA, invalidRecords)
+
+  for (const val of validRecords) {
+    it('converts to BIND format', async function () {
+      const r = new SOA(val).toBind()
+      if (process.env.DEBUG) console.dir(r)
+      assert.strictEqual(r, `$TTL    3600
+$ORIGIN example.com.
+example.com.   IN  SOA ns1.example.com.    matt.example.com. (
+          1
+          7200
+          3600
+          1209600
+          3600
+          )\n\n`)
+    })
+
+    it('converts to tinydns format', async function () {
+      const r = new SOA(val).toTinydns()
+      if (process.env.DEBUG) console.dir(r)
+      assert.strictEqual(r, 'Zexample.com:ns1.example.com.:matt.example.com.:1:7200:3600:1209600:3600:3600::\n')
+    })
+  }
 })