+++ CHANGES.md
@@ -1,6 +1,10 @@
 
 #### 1.N.N - YYYY-MM-DD
 
+- refactored classes into separate files
+- TXT: add toBIND and toTinydns
+- SOA: add toBIND and toTinydns, add setMinimum
+
 
 #### 0.2.2 - 2021-10-20